example: HttpTransport mocking
A illustrative example of the value of interfaces in creating testable software.

In this example, the `http.DefaultClient.Transport` `http.RoundTripper` interface is replaced with a function which returns our bytes as the body and our status code, regardless of the URL the public or private code is trying to call.

When the client utilizes the `http.DefaultClient` we can then test fully the integration of the HTTP request and response handling in our code.