Train February 2024
This is the set of feature shipping in the February departure.

## VideoUpload Receiver Stack

# TODO: This becomes "homewatch-agent"
This contains the configuration code to setup the video upload receiver.

# Features

## --v2 Version2 Flag

This creates a simple `--v2` flag to enable the Version2 api.

## Version2 Watch Paths

This adds the `--v2-watch-paths` flag to allow simple, non-recursive, path watching.

## Watch child paths on startup

When each path in `--v2-watch-paths` is walked the child paths which are directories are added to the FSNotify watcher.

ultralist:27

## Should handle the file write

The file write should have a handler outside the main listener.

ultralist:32

## Add paths on startup and create

On startup, all `--v2-watched-paths` are desceneded through and directories are added.

When a `CREATE` event is received and the path is a directory it is added to the Watcher.

ultralist:27
ultralist:28

## Upload DAV to s3 immediately

This uploads created DAVs to S3 immediately.

ultralist:40

## Publish a metric on create DAV

Adds a new `--v2-enable-metrics` flag. When set an HTTP server is run on `:2112`. It serves the `videos_captured` metric on `/metrics` for Prometheus to scrape.

ultralist:41

## chore: Readme and playbook updates

## Publish metric per camera DAV created

Updates the metric counter for created DAVs including the CameraName as a dimension of the metric.

ultralist:42

## Publish metric per camera DAV upload

Publishes a metric each time a camera uploads a DAV.

ultralist:43

## Track which paths are watched

Track the paths which are watched for the lifetime of execution.

ultralist:30